+++ bta/config.py
@@ -274,7 +274,7 @@ def get_config():
     parser.add_argument("--use_graph", action='store_true', default=False, help=" by default False, use temporal graph.")
     parser.add_argument("--mix_id", type=int,
                         default=0, help="choose 0 to use mixer, 1 to use hyper, 2 to use attention")
-    parser.add_argument("--token_factor", type=int,
+    parser.add_argument("--token_factor", type=float,
                         default=2, help="default is 1")
     parser.add_argument('--max_edges', type=int, default=10)
     parser.add_argument('--num_layers', type=int, default=1)
